# packages/server/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Letter Forge Engine API",
    "description": "API Documentation for Letter Forge Engine",
    "version": "1.0.0"
  },
  "host": "localhost:8888",
  "basePath": "/letter-forge/v1",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT bearer token (format: \"Bearer <token>\")"
    }
  },
...
  "paths": {
...
    "/games/{gameId}/leaderboard": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Submit game score to leaderboard",
        "description": "Submits the final game score to the leaderboard",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "period": {
                  "type": "string",
                  "enum": [
                    "daily",
                    "weekly",
                    "all_time"
                  ],
                  "default": "all_time"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/scores": {
      "post": {
        "tags": [
          "scores"
        ],
        "summary": "Create a new score",
        "description": "Creates a score entry for a game",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string"
                },
                "gameId": {
                  "type": "string"
                },
                "mode": {
                  "type": "string",
                  "enum": [
                    "normal_mode",
                    "time_attack",
                    "survival_mode",
                    "chain_mode"
                  ]
                },
                "points": {
                  "type": "integer"
                }
              },
              "required": [
                "userId",
                "gameId",
                "mode",
                "points"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/scores/game/{gameId}": {
      "get": {
        "tags": [
          "scores"
        ],
        "summary": "Get scores by game ID",
        "description": "Returns all scores for a specific game",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/scores/user/{userId}": {
      "get": {
        "tags": [
          "scores"
        ],
        "summary": "Get scores by user ID",
        "description": "Returns all scores for a specific user",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/scores/user/{userId}/total": {
      "get": {
        "tags": [
          "scores"
        ],
        "summary": "Get total score by user",
        "description": "Returns total score and count for a user",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "mode",
            "in": "query",
            "required": true,
            "type": "string",
            "enum": [
              "normal_mode",
              "time_attack",
              "survival_mode",
              "chain_mode"
            ]
          },
          {
            "name": "period",
            "in": "query",
            "required": true,
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "all_time"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/leaderboard": {
      "get": {
        "tags": [
          "leaderboard"
        ],
        "summary": "Get leaderboard rankings",
        "description": "Returns paginated leaderboard rankings",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": true,
            "type": "string",
            "enum": [
              "normal_mode",
              "time_attack",
              "survival_mode",
              "chain_mode"
            ]
          },
          {
            "name": "period",
            "in": "query",
            "required": true,
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "all_time"
            ]
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "default": 10
          },
          {
            "name": "offset",
            "in": "query",
            "type": "integer",
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/leaderboard/submit": {
      "post": {
        "tags": [
          "leaderboard"
        ],
        "summary": "Submit game score",
        "description": "Submit or update a user's score in the leaderboard",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string"
                },
                "mode": {
                  "type": "string",
                  "enum": [
                    "normal_mode",
                    "time_attack",
                    "survival_mode",
                    "chain_mode"
                  ]
                },
                "period": {
                  "type": "string",
                  "enum": [
                    "daily",
                    "weekly",
                    "all_time"
                  ]
                },
                "score": {
                  "type": "integer"
                }
              },
              "required": [
                "userId",
                "mode",
                "period",
                "score"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/leaderboard/rank/{userId}": {
      "get": {
        "tags": [
          "leaderboard"
        ],
        "summary": "Get user's rank",
        "description": "Returns user's rank in a specific leaderboard",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "mode",
            "in": "query",
            "required": true,
            "type": "string",
            "enum": [
              "normal_mode",
              "time_attack",
              "survival_mode",
              "chain_mode"
            ]
          },
          {
            "name": "period",
            "in": "query",
            "required": true,
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "all_time"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "User not found in leaderboard"
          }
        }
      }
    },
    "/leaderboard/all": {
      "get": {
        "tags": [
          "leaderboard"
        ],
        "summary": "Get all leaderboard data",
        "description": "Returns all leaderboard rankings for all modes and periods",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/milestones/{milestoneId}": {
      "get": {
        "tags": [
          "milestones"
        ],
        "summary": "Get milestone by ID",
        "description": "Returns a specific milestone",
        "parameters": [
          {
            "name": "milestoneId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Milestone not found"
          }
        }
      }
    },
    "/milestones/user/{userId}": {
      "get": {
        "tags": [
          "milestones"
        ],
        "summary": "Get user milestones",
        "description": "Returns completed milestones for a user",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/milestones/check/{userId}": {
      "post": {
        "tags": [
          "milestones"
        ],
        "summary": "Check and unlock milestones",
        "description": "Computes real stats from completed games and unlocks any achieved milestones",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Ignored - stats are computed server-side",
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}